+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Update makerDAO dai branding [#1668](https://github.com/MyEtherWallet/MyEtherWallet/pull/1668)
 - Update translations[#1651](https://github.com/MyEtherWallet/MyEtherWallet/pull/1651)
 - Display token symbol [#1674](https://github.com/MyEtherWallet/MyEtherWallet/pull/1674)
+- Refactor translations [#1682](https://github.com/MyEtherWallet/MyEtherWallet/pull/1682)
 
 ### Feature