+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Feature
 
-- Add MEWconnect footer link [#1675] (https://github.com/MyEtherWallet/MyEtherWallet/pull/1675)
+- Add MEWconnect footer link [#1675](https://github.com/MyEtherWallet/MyEtherWallet/pull/1675)
 
 ### Release v5.2.4